LPD-63562 Move classes to make them testable
