+++ modules/apps/depot/depot-api/src/main/java/com/liferay/depot/item/selector/DepotGroupItemSelectorCriterion.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.depot.web.internal.item.selector;
+package com.liferay.depot.item.selector;
 
 import com.liferay.depot.constants.DepotConstants;
 import com.liferay.item.selector.criteria.group.criterion.GroupItemSelectorCriterion;
