+++ modules/apps/depot/depot-api/src/main/java/com/liferay/depot/search/DepotEntrySearch.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.depot.web.internal.search;
+package com.liferay.depot.search;
 
 import com.liferay.depot.model.DepotEntry;
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
