+++ modules/apps/depot/depot-api/src/main/resources/com/liferay/depot/item/selector/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
