+++ modules/apps/depot/depot-api/src/main/resources/com/liferay/depot/search/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
